app/requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
click==8.0.4
colorama==0.4.4
databases==0.5.5
fastapi==0.75.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
watchgod==0.8.1
websockets==10.2
